requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0